% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automap.R
\name{automap}
\alias{automap}
\title{Automatically create 2D and 3D maps using USGS and ESRI map data}
\usage{
automap(
  lat = NULL,
  lng = NULL,
  distance = 10,
  method = "2d",
  img.width = 600,
  img.height = 600,
  overlay = NULL,
  z = 9,
  overlay.alpha = 0.75,
  colorscale = "imhof4",
  color.intensity = 1,
  max.darken = 0.5,
  sun.angle = 315,
  sun.altitude = 45,
  water.cutoff = 0.999,
  water.min.area = NULL,
  water.max.height = NULL,
  solid = TRUE,
  shadow = TRUE,
  water = FALSE,
  waterdepth = 0,
  theta = 45,
  phi = 45,
  fov = 0,
  zoom = 1,
  dist.unit = "km",
  coord.unit = "degrees",
  heightmap = NULL,
  texture = NULL
)
}
\arguments{
\item{lat}{The latitude for the map's centroid, in decimal degrees or
radians.}

\item{lng}{The longitude for the map's centroid, in decimal degrees or
radians.}

\item{distance}{The distance between the centroid and any corner of the
(square) output map to include.}

\item{method}{Should a 2d or 3d plot be produced?}

\item{img.width}{Image width, in pixels}

\item{img.height}{Image height, in pixels}

\item{overlay}{ESRI overlay map to include, if any -- see
\code{\link[spacey]{get_image_overlay}} for list of options. If specifying
a local file with \code{from.file}, any non \code{NULL} value will add your
local file as an overlay.}

\item{z}{zscale, passed to various rayshader functions. Defined as
the ratio between the x and y spacing (which are assumed to be equal)
and the z axis. For contiguous United States, USGS data is generally
available at a ~9 meter spacing, with elevation provided in meter increments,
resulting in a z value of 9 returning roughly representative maps; decrease
this value to exaggerate elevation features.}

\item{overlay.alpha}{Alpha value for the optional overlay layer.}

\item{colorscale}{Color scale for land and water elements. If a vector of
length 1, the same color scale will be applied for both land and water. If
greater than  length 1, values named "water" or "watercolor" will be used
for water coloring, while "land" or "landcolor" will be used for land.}

\item{color.intensity}{Intensity of color mapping -- higher values result in
more intense colors.}

\item{max.darken}{Passed to \code{\link[rayshader]{add_shadow}}. The lower
limit for how much the image will be darkened. 0 is completely black, 1 means
the shadow map will have no effect.}

\item{sun.angle}{Angle around the matrix from which lights originate. Values
line up with compass directions -- so 0 is directly North, while the default
315 places the sun in the Northwest.}

\item{sun.altitude}{Angle in degrees from horizon from which light
originates. Bounded [0, 90].}

\item{water.cutoff}{Passed to \code{\link[rayshader]{detect_water}}. Defined
therein as the lower limit of the z-component of the unit normal vector to
be classified as water.}

\item{water.min.area}{Passed to \code{\link[rayshader]{detect_water}}. The
minimum possible area to consider a body of water.}

\item{water.max.height}{Passed to \code{\link[rayshader]{detect_water}}. If
provided, the maximum height a point can be classified water.}

\item{solid}{Logical -- should the output be rendered as a solid
(\code{TRUE}) or just a surface (\code{FALSE})?}

\item{shadow}{Logical -- should shadows be rendered?}

\item{water}{Logical -- should water be rendered?}

\item{waterdepth}{Water level.}

\item{theta}{Rotation around z axis.}

\item{phi}{Azimuth angle.}

\item{fov}{Field of view angle.}

\item{zoom}{Zoom factor.}

\item{dist.unit}{Units for the distance argument. All units are converted to
kilometers, with some errors in the transition due to floating point
arithmetic -- so if high accuracy is needed, convert units to kilometers
beforehand.}

\item{coord.unit}{Units for latitude and longitude, in either decimal degrees
or radians.}

\item{heightmap}{Heightmap array, either provided by the user or pulled
via the USGS national map API.}

\item{texture}{Overlay image array, either provided by the user or pulled
from the ESRI map texture API.}
}
\value{
A rayshader map object, by default printed and returned invisibly.
}
\description{
This function takes a latitude/longitude coordinate pair alongside a desired
distance to map and retrieves USGS (and optionally ESRI) map data, converts
said data into matrices, and runs the data through \code{rayshader} using
sensible defaults in order to return a 2D shaded map relief. It requires a
functioning internet connection in order to retrieve data.
}
\examples{
\dontrun{
automap(44.121268, -73.903734)
automap(44.121268, -73.903734, overlay = "World_Imagery")
automap(44.121268, -73.903734, overlay = "World_Imagery", method = "3d")
}

}
\references{
Archuleta, C.M., Constance, E.W., Arundel, S.T., Lowe, A.J., Mantey, K.S.,
and Phillips, L.A., 2017, The National Map seamless digital elevation model
specifications: U.S. Geological Survey Techniques and Methods, book 11,
chap. B9, 39 p., https://doi.org/10.3133/tm11B9
}
