% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automap_2d.R
\name{automap_2d}
\alias{automap_2d}
\title{Automatically create 2D maps using USGS and ESRI map data}
\usage{
automap_2d(
  lat,
  lng,
  distance = 10,
  major.dim = 600,
  overlay = NULL,
  z = 30,
  overlay.alpha = 0.75,
  colorscale = "imhof4",
  max.darken = 0.5,
  save.tif = FALSE,
  save.png = FALSE,
  tif.filename = NULL,
  png.filename = NULL,
  dist.unit = c("km", "miles", "m", "ft"),
  coord.unit = c("degrees", "radians"),
  sr_bbox = 4326,
  sr_image = 4326,
  print.map = TRUE
)
}
\arguments{
\item{lat}{The latitude for the map's centroid, in decimal degrees or
radians.}

\item{lng}{The longitude for the map's centroid, in decimal degrees or
radians.}

\item{distance}{The distance between the centroid and any corner of the
(square) output map to include.}

\item{major.dim}{Image size in the major direction -- as this function
produces square maps, this is equivalent to pixels along any edge.}

\item{overlay}{ESRI overlay map to include, if any -- see
\code{\link[spacey]{get_image_overlay}} for list of options.}

\item{z}{zscale, passed to various rayshader functions. Defined as
the ratio between the x and y spacing (which are assumed to be equal)
and the z axis.}

\item{overlay.alpha}{Alpha value for the optional overlay layer.}

\item{colorscale}{Color scale for land and water elements. If a vector of
length 1, the same color scale will be applied for both land and water. If
greater than  length 1, values named "water" or "watercolor" will be used
for water coloring, while "land" or "landcolor" will be used for land.}

\item{max.darken}{Passed to \code{\link[rayshader]{add_shadow}}.}

\item{save.tif}{Logical -- should the height map be saved?}

\item{save.png}{Logical -- should the overlay image be saved?}

\item{tif.filename}{Filename for the heightmap if \code{save.tif} is TRUE.}

\item{png.filename}{Filename for the overlay if \code{save.png} is TRUE.}

\item{dist.unit}{Units for the distance argument. All units are converted to
kilometers, with some errors in the transition due to floating point
arithmetic -- so if high accuracy is needed, convert units to kilometers
beforehand.}

\item{coord.unit}{Units for latitude and longitude, in either decimal degrees
or radians.}

\item{sr_bbox}{Spatial reference code (ISO 19111) for bounding box}

\item{sr_image}{Spatial reference code (ISO 19111) for image}

\item{print.map}{Logical -- should the output map be printed?}
}
\value{
A rayshader map object, by default printed and returned invisibly.
}
\description{
This function takes a latitude/longitude coordinate pair alongside a desired
distance to map and retrieves USGS (and optionally ESRI) map data, converts
said data into matrices, and runs the data through \code{rayshader} using
sensible defaults in order to return a 2D shaded map relief. It requires a
functioning internet connection in order to retrieve data.
}
