% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert degrees to radians}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{Value in degrees to be converted}
}
\value{
Value in radians
}
\description{
Convert degrees to radians
}
\keyword{internal}
