% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract latitude and longitude from a provided object}
\usage{
extract_coords(coords)
}
\arguments{
\item{coords}{An object to try and extract coordinates from}
}
\value{
A vector of length 2 containing object latitude and longitude
}
\description{
Extract latitude and longitude from a provided object
}
\keyword{internal}
