% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounding_box.R
\name{get_coord_bounding_box}
\alias{get_coord_bounding_box}
\title{Get bounding box for set of coordinate points}
\usage{
get_coord_bounding_box(lat, lng)
}
\arguments{
\item{lat}{A vector of latitudes}

\item{lng}{A vector of longitudes}
}
\value{
A list of length 2, containing the bottom-left (named "bl") and
top-right (named "tr") coordinates of the bounding box.
}
\description{
Get bounding box for set of coordinate points
}
