% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_img_size}
\alias{get_img_size}
\title{Calculate image height and width from a bounding box and edge length}
\usage{
get_img_size(bbox, lat = NULL, lng = NULL, major.dim = 600)
}
\arguments{
\item{bbox}{The bounding box to calculate image size for}

\item{lat}{A quoted string indicating what named value in the bounding box
represents latitude. If NULL, will be inferred from bounding box names.}

\item{lng}{A quoted string indicating what named value in the bounding box
representes longitude. If NULL, will be inferred from bounding box names.}

\item{major.dim}{The length of the (rectangular) image's largest side.}
}
\value{
A length-2 vector of image height and width
}
\description{
Calculate image height and width from a bounding box and edge length
}
\keyword{internal}
