% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_texture.R
\name{get_texture}
\alias{get_texture}
\title{Get color textures for a map}
\usage{
get_texture(colorscale)
}
\arguments{
\item{colorscale}{The color scale to return an RGB array for}
}
\value{
A matrix object provided by \code{\link[png]{readPNG}}, suitable for
use with \code{rayshader} and similar mapping utilities.
}
\description{
This function implements a number of new color textures for rayshader maps,
allowing users to reference any of the "spacey" palettes without needing to
create arrays themselves. It is extremely unlikely users will need to call
this function themselves; instead, most users will call it via
\code{[automap]}.
}
\examples{
get_texture("spacey1")
}
\keyword{internal}
