% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automap.R
\name{make_copy}
\alias{make_copy}
\title{Make a copy}
\usage{
make_copy(map_object)
}
\arguments{
\item{map_object}{Your \code{spacey_map} object.}
}
\description{
This function is an extremely thin wrapper around the R6 \code{clone} method
to allow for more natural R code.
}
