% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert radians to degrees}
\usage{
rad_to_deg(rad)
}
\arguments{
\item{rad}{Value in radians to be converted}
}
\value{
Value in degrees
}
\description{
Convert radians to degrees
}
\keyword{internal}
