% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automap.R
\name{spacey_map}
\alias{spacey_map}
\title{spacey map objects}
\description{
This class powers \code{automap} and related functions.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lat}}{The latitude for the map's centroid, in decimal degrees or
radians.}

\item{\code{lng}}{The longitude for the map's centroid, in decimal degrees or
radians.}

\item{\code{distance}}{The distance between the centroid and any corner of the
(square) output map to include.}

\item{\code{method}}{Should a 2d or 3d plot be produced?}

\item{\code{img.width}}{Image width, in pixels}

\item{\code{img.height}}{Image height, in pixels}

\item{\code{overlay}}{ESRI overlay map to include, if any -- see
\code{\link[spacey]{get_image_overlay}} for list of options. If specifying
a local file with \code{from.file}, any non \code{NULL} value will add your
local file as an overlay.}

\item{\code{z}}{zscale, passed to various rayshader functions. Defined as
the ratio between the x and y spacing (which are assumed to be equal)
and the z axis. For contiguous United States, USGS data is generally
available at a ~9 meter spacing, with elevation provided in meter increments,
resulting in a z value of 9 returning roughly representative maps; decrease
this value to exaggerate elevation features.}

\item{\code{land.z}}{zscale for land elements.}

\item{\code{water.z}}{zscale for water elements.}

\item{\code{overlay.alpha}}{Alpha value for the optional overlay layer.}

\item{\code{colorscale}}{Color scale for land and water elements. If a vector of
length 1, the same color scale will be applied for both land and water. If
greater than  length 1, values named "water" will be used
for water coloring, while "land" will be used for land.}

\item{\code{water.color}}{Color scale for water elements}

\item{\code{land.color}}{Color scale for land elements}

\item{\code{color.intensity}}{Intensity of color mapping -- higher values result in
more intense colors.}

\item{\code{max.darken}}{Passed to \code{\link[rayshader]{add_shadow}}. The lower
limit for how much the image will be darkened. 0 is completely black, 1 means
the shadow map will have no effect.}

\item{\code{sun.angle}}{Angle around the matrix from which lights originate. Values
line up with compass directions -- so 0 is directly North, while the default
315 places the sun in the Northwest.}

\item{\code{sun.altitude}}{Angle in degrees from horizon from which light
originates. Bounded [0, 90].}

\item{\code{water.cutoff}}{Passed to \code{\link[rayshader]{detect_water}}. Defined
therein as the lower limit of the z-component of the unit normal vector to
be classified as water.}

\item{\code{water.min.area}}{Passed to \code{\link[rayshader]{detect_water}}. The
minimum possible area to consider a body of water.}

\item{\code{water.max.height}}{Passed to \code{\link[rayshader]{detect_water}}. If
provided, the maximum height a point can be classified water.}

\item{\code{solid}}{Logical -- should the output be rendered as a solid
(\code{TRUE}) or just a surface (\code{FALSE})?}

\item{\code{shadow}}{Logical -- should shadows be rendered?}

\item{\code{water}}{Logical -- should water be rendered?}

\item{\code{waterdepth}}{Water level.}

\item{\code{theta}}{Rotation around z axis.}

\item{\code{phi}}{Azimuth angle.}

\item{\code{fov}}{Field of view angle.}

\item{\code{zoom}}{Zoom factor for 3D maps.}

\item{\code{dist.unit}}{Units for the distance argument. All units are converted to
kilometers, with some errors in the transition due to floating point
arithmetic -- so if high accuracy is needed, convert units to kilometers
beforehand.}

\item{\code{coord.unit}}{Units for latitude and longitude, in either decimal degrees
or radians.}

\item{\code{bbox}}{Bounding box for map data, calculated based on lat/lng and
distance values.}

\item{\code{heightmap}}{Heightmap array, either provided by the user or pulled
via the USGS national map API.}

\item{\code{texture}}{Overlay image array, either provided by the user or pulled
from the ESRI map texture API.}

\item{\code{normals}}{Normal vectors, calculated and cached to speed
recalculation.}

\item{\code{water_future}}{Array of water values.}

\item{\code{base_future}}{Array of base elevation sphere shading.}

\item{\code{rayshade_future}}{Array of ray shaded values.}

\item{\code{ambient_future}}{Array of ambient shadow shading values.}

\item{\code{map_obj}}{The final map image.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{spacey_map$new()}}
\item \href{#method-print}{\code{spacey_map$print()}}
\item \href{#method-update_values}{\code{spacey_map$update_values()}}
\item \href{#method-clone}{\code{spacey_map$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new spacey_map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spacey_map$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Fields passed from \code{automap} to initalize the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{spacey_map} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print out the map object within a spacey_map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spacey_map$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_values"></a>}}
\if{latex}{\out{\hypertarget{method-update_values}{}}}
\subsection{Method \code{update_values()}}{
Update values within the map object and recalculate arrays as necessary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spacey_map$update_values(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any combination of fields accepted by \code{spacey_map}.
Fields must be named.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spacey_map$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
