% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automap.R
\name{update_values}
\alias{update_values}
\title{Update values within the map object and recalculate arrays as necessary}
\usage{
update_values(map_object, ...)
}
\arguments{
\item{map_object}{Your \code{spacey_map} object.}

\item{...}{Any combination of fields accepted by \code{\link[spacey]{automap}}.
Fields must be named.}
}
\description{
This function will update the values within whatever object you pass it;
assigning it to another name will create a new reference (but not a new
object) and both names will point to the same, modified object. To create a
new object, use \code{\link[spacey]{make_copy}} or your object's \code{clone}
method.
}
